epvm scheduler

+++ task/task.go
@@ -14,8 +14,8 @@ type Task struct {
 	Name          string
 	State         State
 	Image         string
-	Memory        int
-	Disk          int
+	Memory        int64
+	Disk          int64
 	ExposedPorts  nat.PortSet
 	HostPorts     nat.PortMap
 	PortBindings  map[string]string
